+++ shortcodes/stable_diffusion/controlnet.py
@@ -1,3 +1,4 @@
+from modules.shared import cmd_opts
 class Shortcode():
 	def __init__(self,Unprompted):
 		self.Unprompted = Unprompted
@@ -5,20 +6,30 @@ class Shortcode():
 		self.detect_resolution = 512
 		self.save_memory = False
 		self.steps = 20
+		self.can_run = False
+		self.temp_no_half = False
 	
 	def run_atomic(self, pargs, kwargs, context):
 		if "init_images" not in self.Unprompted.shortcode_user_vars:
 			self.Unprompted.log("This shortcode is only supported in img2img mode.","ERROR")
+
 		
-		self.steps = self.Unprompted.shortcode_user_vars["steps"]
 		# Hacky way of bypassing the normal img2img routine
+		self.steps = self.Unprompted.shortcode_user_vars["steps"]
 		self.Unprompted.shortcode_user_vars["steps"] = 1
+		self.can_run = True
+
+		# Workaround for half precision error
+		self.temp_no_half = cmd_opts.no_half
+		cmd_opts.no_half = True
 
 		if "detect_resolution" in kwargs: self.detect_resolution = kwargs["detect_resolution"]
 		if "save_memory" in pargs: self.save_memory = True
 		return("")
 
 	def after(self,p=None,processed=None):
+		if "init_images" not in self.Unprompted.shortcode_user_vars:
+			self.Unprompted.log("This shortcode is only supported in img2img mode.","ERROR")
 		import torch
 		import cv2
 		import einops
@@ -27,22 +38,25 @@ class Shortcode():
 		from PIL import Image
 
 		from modules.images import flatten
-		from modules.shared import opts	
+		from modules.shared import opts, sd_model
 
 		from lib_unprompted.stable_diffusion.controlnet.annotator.util import resize_image, HWC3
 		from lib_unprompted.stable_diffusion.controlnet.annotator.openpose import apply_openpose
 		from lib_unprompted.stable_diffusion.controlnet.cldm.model import create_model, load_state_dict
 		from lib_unprompted.stable_diffusion.controlnet.ldm.models.diffusion.ddim import DDIMSampler
 
+		# from modules import sd_models
+
 		import sys
 		sys.path.append(f"{self.Unprompted.base_dir}/lib_unprompted/stable_diffusion/controlnet")
 
-		model = create_model(f"{self.Unprompted.base_dir}/lib_unprompted/stable_diffusion/controlnet/models/cldm_v15.yaml").cpu()
-		model.load_state_dict(load_state_dict(f"{self.Unprompted.base_dir}/lib_unprompted/stable_diffusion/controlnet/models/control_sd15_openpose.pth", location='cpu'),strict=False)
-		model = model.cuda()
-		ddim_sampler = DDIMSampler(model)
+		from modules import sd_models
+		info = sd_models.get_closet_checkpoint_match("control_sd15_openpose")
+		if (info): sd_models.reload_model_weights(None,info)
+		sd_model = sd_model.cuda()
+		ddim_sampler = DDIMSampler(sd_model)
 
-		image_resolution = 512
+		image_resolution = min(self.Unprompted.shortcode_user_vars["width"],self.Unprompted.shortcode_user_vars["height"])
 
 		num_samples = self.Unprompted.shortcode_user_vars["batch_size"]
 
@@ -62,20 +76,15 @@ class Shortcode():
 				control = torch.stack([control for _ in range(num_samples)], dim=0)
 				control = einops.rearrange(control, 'b h w c -> b c h w').clone()
 
-				# we're already seeded bro 8)
-				#if seed == -1:
-				#	seed = random.randint(0, 65535)
-				# seed_everything(seed)
-
 				if self.save_memory:
-					model.low_vram_shift(is_diffusing=False)
+					sd_model.low_vram_shift(is_diffusing=False)
 
-				cond = {"c_concat": [control], "c_crossattn": [model.get_learned_conditioning([self.Unprompted.shortcode_user_vars["prompt"]] * num_samples)]}
-				un_cond = {"c_concat": [control], "c_crossattn": [model.get_learned_conditioning([self.Unprompted.shortcode_user_vars["negative_prompt"]] * num_samples)]}
+				cond = {"c_concat": [control], "c_crossattn": [sd_model.get_learned_conditioning([self.Unprompted.shortcode_user_vars["prompt"]] * num_samples)]}
+				un_cond = {"c_concat": [control], "c_crossattn": [sd_model.get_learned_conditioning([self.Unprompted.shortcode_user_vars["negative_prompt"]] * num_samples)]}
 				shape = (4, H // 8, W // 8)
 
 				if self.save_memory:
-					model.low_vram_shift(is_diffusing=True)
+					sd_model.low_vram_shift(is_diffusing=True)
 
 				samples, intermediates = ddim_sampler.sample(self.steps, num_samples,
 															shape, cond, verbose=False, eta=self.Unprompted.shortcode_user_vars["denoising_strength"],
@@ -83,9 +92,9 @@ class Shortcode():
 															unconditional_conditioning=un_cond)
 
 				if self.save_memory:
-					model.low_vram_shift(is_diffusing=False)
+					sd_model.low_vram_shift(is_diffusing=False)
 
-				x_samples = model.decode_first_stage(samples)
+				x_samples = sd_model.decode_first_stage(samples)
 				x_samples = (einops.rearrange(x_samples, 'b c h w -> b h w c') * 127.5 + 127.5).cpu().numpy().clip(0, 255).astype(np.uint8)
 
 				for i in range(num_samples):
@@ -93,11 +102,15 @@ class Shortcode():
 					output_map = Image.fromarray(detected_map)
 					processed.images.append(output)
 					processed.images.append(output_map)
-		
-		print("done")
+
+		# Set back to user defined value
+		cmd_opts.no_half = self.temp_no_half
 		
 		# processed.images.append([detected_map])
 		return(processed)
+
+	def cleanup(self):
+		self.save_memory = False
 
 	def ui(self,gr):
 		pass